Refactor double-fetch for token changes
Previously we would perform a `getIdTokenResult()`, followed by a
`getIdToken()` call. This is two API calls to Firebase, and can be
simplfied by using the `token` property from the resutl of
`getIdTokenResult()`.

Also updates the `onIdTokenChanged` flow to handle errors more
gracefully, by `return`ing early if we are unable to retrieve the token
details (rather than setting `null` data to memory.

Observing the source of `getIdTokenResult` at
https://github.com/firebase/firebase-js-sdk/blob/6d31930b3abe1588ae81a5c14b59cd386fddc718/packages/auth/src/core/user/id_token_result.ts#L55-L88
it is evident that it calls `getIdToken` internally.