341 Change membership endpoint to allow client to select membership type
Update routes/schemas.

Change to a POST request instead of GET.

Changed route testing to include membership type.

Created a new requestBody for the payments controller. 

Removed PricingService and the tests for it.

Updated client frontend for changes as per the ticket.

Require another ticket to remove hard coded membershipType in frontend.