Neaten endpoint tests draft
Just to check and see if everything will still work.

Created a `setupTests.ts` file to include most of the generic code that's repeated or unecessary in most of the files for the future.

Also added the setup file to `jest.config.server.ts`

Removed redundant setup code from `routes.test.ts`